+++ packages/lumx-react/src/components/mosaic/Mosaic.tsx
@@ -1,21 +1,12 @@
 import React from 'react';
 
-import { AspectRatio, CrossOrigin, FocusPoint, Theme, Thumbnail } from '@lumx/react';
+import { AspectRatio, Theme, Thumbnail, ThumbnailProps } from '@lumx/react';
 import { COMPONENT_PREFIX } from '@lumx/react/constants';
 import { GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
 import classNames from 'classnames';
 import take from 'lodash/take';
 
-interface MosaicElement {
-    /**
-     * Allows images that are loaded from foreign origins
-     * to be used as if they had been loaded from the current origin.
-     */
-    crossOrigin?: CrossOrigin;
-    focusPoint?: FocusPoint;
-    /** Active cross origin. */
-    isCrossOriginEnabled?: boolean;
-    url: string;
+interface MosaicElement extends Exclude<ThumbnailProps, 'aspectRatio' | 'fillHeight'> {
     onClick?(index: number): void;
 }
 
@@ -66,14 +57,10 @@ const Mosaic: React.FC<MosaicProps> = ({ className, theme = DEFAULT_PROPS.theme,
                 return (
                     <div key={index} className={`${CLASSNAME}__thumbnail`}>
                         <Thumbnail
+                            {...thumbnail}
                             aspectRatio={AspectRatio.free}
-                            crossOrigin={thumbnail.crossOrigin}
                             fillHeight
-                            focusPoint={thumbnail.focusPoint}
-                            image={thumbnail.url}
-                            isCrossOriginEnabled={thumbnail.isCrossOriginEnabled}
                             tabIndex={thumbnail.onClick && '0'}
-                            theme={theme}
                             onClick={handleClick}
                         />